ACTIN-808: move builder to test factory
